requirements.txt
praw
discord.py
wikipedia
asyncio
python-dotenv
